# AquaSense/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "AquaSense API",
    "version": "1.0.0",
    "description": "API для керування AquaSense системою"
  },
  "host": "localhost:3000",
  "schemes": ["http"],
...
  "paths": {
...
    "/aquariums": {
      "post": {
        "tags": ["Aquariums manage"],
        "summary": "Створити новий акваріум",
        "description": "Додає новий акваріум до бази даних",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "description": "Дані нового акваріума",
            "schema": {
              "type": "object",
              "properties": {
                "user_id": { "type": "integer" },
                "name": { "type": "string" },
                "specification": { "type": "string" }
              },
              "required": ["user_id", "name"]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Акваріум успішно створено"
          },
          "500": {
            "description": "Помилка створення акваріума"
          }
        }
      },
      "get": {
        "tags": ["Aquariums manage"],
        "summary": "Отримати всі акваріуми",
        "description": "Повертає список усіх акваріумів у системі",
        "responses": {
          "200": {
            "description": "Успішно",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": { "type": "integer" },
                  "user_id": { "type": "integer" },
                  "name": { "type": "string" },
                  "specification": { "type": "string" },
                  "created_at": { "type": "string", "format": "date-time" }
                }
              }
            }
          }
        }
      }
    },
    "/aquariums/user/{userId}": {
      "get": {
        "tags": ["Aquariums manage"],
        "summary": "Отримати всі акваріуми для користувача за його id",
        "description": "Повертає список акваріумів, що належать конкретному користувачу за його id",
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "type": "integer",
            "description": "id користувача"
          }
        ],
        "responses": {
          "200": {
            "description": "Успішно",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": { "type": "integer" },
                  "user_id": { "type": "integer" },
                  "name": { "type": "string" },
                  "specification": { "type": "string" },
                  "created_at": { "type": "string", "format": "date-time" }
                }
              }
            }
          },
          "404": {
            "description": "Користувача не знайдено або у користувача немає акваріумів"
          }
        }
      }
    },
    "/aquariums/{id}": {
      "get": {
        "tags": ["Aquariums manage"],
        "summary": "Отримати акваріум за його id",
        "description": "Повертає інформацію про акваріум за його id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "description": "id акваріума"
          }
        ],
        "responses": {
          "200": {
            "description": "Успішно",
            "schema": {
              "type": "object",
              "properties": {
                "id": { "type": "integer" },
                "user_id": { "type": "integer" },
                "name": { "type": "string" },
                "specification": { "type": "string" },
                "created_at": { "type": "string", "format": "date-time" }
              }
            }
          },
          "404": {
            "description": "Акваріум не знайдено"
          }
        }
      },
      "put": {
        "tags": ["Aquariums manage"],
        "summary": "Оновити інформацію про акваріум",
        "description": "Оновлює інформацію про акваріум за його id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "description": "id акваріума"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "description": "Нові дані акваріума",
            "schema": {
              "type": "object",
              "properties": {
                "name": { "type": "string" },
                "specification": { "type": "string" }
              },
              "required": ["name", "specification"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Акваріум успішно оновлено"
          },
          "404": {
            "description": "Акваріум не знайдено"
          }
        }
      },
      "delete": {
        "tags": ["Aquariums manage"],
        "summary": "Видалити акваріум",
        "description": "Видаляє акваріум за його id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "description": "id акваріума"
          }
        ],
        "responses": {
          "200": {
            "description": "Акваріум успішно видалено"
          },
          "404": {
            "description": "Акваріум не знайдено"
          }
        }
      }
    },
    "/devices": {
      "post": {
        "tags": ["Devices manager"],
        "summary": "Створити новий девайс",
        "description": "Додає новий девайс до акваріума",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "description": "Дані нового девайса",
            "schema": {
              "type": "object",
              "properties": {
                "aquarium_id": { "type": "integer" },
                "name": { "type": "string" },
                "status": { "type": "string", "enum": ["on", "off"] }
              },
              "required": ["aquarium_id", "name"]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Девайс успішно створено"
          },
          "500": {
            "description": "Помилка створення девайса"
          }
        }
      },
      "get": {
        "tags": ["Devices manager"],
        "summary": "Отримати всі девайси",
        "description": "Повертає список усіх девайсів у системі",
        "responses": {
          "200": {
            "description": "Успішно",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": { "type": "integer" },
                  "aquarium_id": { "type": "integer" },
                  "name": { "type": "string" },
                  "status": { "type": "string", "enum": ["on", "off"] },
                  "created_at": { "type": "string", "format": "date-time" }
                }
              }
            }
          }
        }
      }
    },
    "/devices/aquarium/{aquariumId}": {
      "get": {
        "tags": ["Devices manager"],
        "summary": "Отримати всі девайси для акваріуму за його id",
        "description": "Повертає список девайсів, що належать до конкретного акваріума за його id",
        "parameters": [
          {
            "in": "path",
            "name": "aquariumId",
            "required": true,
            "type": "integer",
            "description": "id акваріума"
          }
        ],
        "responses": {
          "200": {
            "description": "Успішно",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": { "type": "integer" },
                  "aquarium_id": { "type": "integer" },
                  "name": { "type": "string" },
                  "status": { "type": "string", "enum": ["on", "off"] },
                  "created_at": { "type": "string", "format": "date-time" }
                }
              }
            }
          },
          "404": {
            "description": "Акваріум не знайдено або в акваріумі немає девайсів"
          }
        }
      }
    },
    "/devices/{id}": {
      "get": {
        "tags": ["Devices manager"],
        "summary": "Отримати девайс за його id",
        "description": "Повертає інформацію про девайс за його id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "description": "id девайса"
          }
        ],
        "responses": {
          "200": {
            "description": "Успішно",
            "schema": {
              "type": "object",
              "properties": {
                "id": { "type": "integer" },
                "aquarium_id": { "type": "integer" },
                "name": { "type": "string" },
                "status": { "type": "string", "enum": ["on", "off"] },
                "created_at": { "type": "string", "format": "date-time" }
              }
            }
          },
          "404": {
            "description": "Девайс не знайдено"
          }
        }
      },
      "put": {
        "tags": ["Devices manager"],
        "summary": "Оновити девайс",
        "description": "Оновлює інформацію про девайс за його id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "description": "id девайса"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "description": "Нові дані девайса",
            "schema": {
              "type": "object",
              "properties": {
                "name": { "type": "string" },
                "status": { "type": "string", "enum": ["on", "off"] }
              },
              "required": ["name", "status"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Девайс успішно оновлено"
          },
          "404": {
            "description": "Девайс не знайдено"
          }
        }
      },
      "delete": {
        "tags": ["Devices manager"],
        "summary": "Видалити девайс",
        "description": "Видаляє девайс за його id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "description": "id девайса"
          }
        ],
        "responses": {
          "200": {
            "description": "Девайс успішно видалено"
          },
          "404": {
            "description": "Девайс не знайдено"
          }
        }
      }
    },
...
    "/sensors/aquarium/{aquariumId}": {
      "get": {
        "tags": ["Sensors manager"],
        "summary": "Отримати всі датчики для акваріуму за його id",
        "description": "Повертає список датчиків, що належать до конкретного акваріума за його id",
        "parameters": [
          {
            "in": "path",
            "name": "aquariumId",
            "required": true,
            "type": "integer",
            "description": "id акваріума"
          }
        ],
        "responses": {
          "200": {
            "description": "Успішно",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": { "type": "integer" },
                  "aquarium_id": { "type": "integer" },
                  "type": { "type": "string" },
                  "value": { "type": "number", "format": "float" },
                  "created_at": { "type": "string", "format": "date-time" }
                }
              }
            }
          },
          "404": {
            "description": "Акваріум не знайдено або в акваріумі немає датчиків"
          }
        }
      }
    },
...
  }
}